Move UnwindReason to common moduel
