+++ lib/vm/src/trap/mod.rs
@@ -12,12 +12,12 @@ mod traphandlers;
 #[path = "traphandlers_baremetal.rs"]
 mod traphandlers;
 
-pub use trap::Trap;
+pub use trap::{Trap, UnwindReason};
 pub use traphandlers::{
     MAX_STACK_SIZE, TrapHandlerFn, VMConfig, catch_traps, drain_stack_pool, get_stack_size,
     on_host_stack, raise_lib_trap, raise_user_trap, set_stack_size, wasmer_call_trampoline,
 };
 #[cfg(feature = "baremetal")]
-pub use traphandlers::{UnwindReason, install_unwinder};
+pub use traphandlers::install_unwinder;
 pub use traphandlers::{init_traps, resume_panic};
 pub use wasmer_types::TrapCode;
